refactor: update generic component filenames to PascalCase
