+++ web/components/BannedIPsTable.tsx
