+++ web/components/LogTable.tsx
