+++ web/components/Offline.tsx
@@ -2,9 +2,9 @@ import { BookTwoTone, MessageTwoTone, PlaySquareTwoTone, ProfileTwoTone } from '
 import { Card, Col, Row, Typography } from 'antd';
 import Link from 'next/link';
 import { useContext } from 'react';
-import LogTable from './log-table';
+import LogTable from './LogTable';
 import OwncastLogo from './common/Logo/Logo';
-import NewsFeed from './news-feed';
+import NewsFeed from './NewsFeed';
 import { ConfigDetails } from '../types/config-section';
 import { ServerStatusContext } from '../utils/server-status-context';
