+++ web/components/StatisticItem.tsx
