+++ web/pages/admin/chat/users.tsx
@@ -8,9 +8,9 @@ import {
   MODERATORS,
   BANNED_IPS,
 } from '../../../utils/apis';
-import UserTable from '../../../components/user-table';
-import ClientTable from '../../../components/client-table';
-import BannedIPsTable from '../../../components/banned-ips-table';
+import UserTable from '../../../components/UserTable';
+import ClientTable from '../../../components/ClientTable';
+import BannedIPsTable from '../../../components/BannedIPsTable';
 
 const { TabPane } = Tabs;
